svg/painting/parsing/stroke-width-valid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="600"><title>SVG Painting: parsing stroke-width with valid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/painting.html#StrokeWidth" rel="help"/><h:meta content="stroke-width supports the full grammar '&lt;length-percentage&gt;' and unitless." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_valid_value("stroke-width", "0", "0px");
test_valid_value("stroke-width", "10", "10px");
test_valid_value("stroke-width", "1px");
test_valid_value("stroke-width", "calc(2em + 3ex)");
test_valid_value("stroke-width", "4%");
test_valid_value("stroke-width", "5vmin");
test_valid_value("stroke-width", "calc(50% + 60px)");

  ]]></script></svg>
